usuario -> corrigindo erros de duplicidade
