--- src/main/java/com/belval/adocaoanimais/controller/LoginController.java
@@ -1,12 +1,0 @@
-package com.belval.adocaoanimais.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LoginController {
-	@GetMapping("/pet/login")
-	public String login() {
-		return "login";
-	}
-}
