--- src/main/java/com/belval/adocaoanimais/controller/RegisterController.java
@@ -1,16 +1,0 @@
-package com.belval.adocaoanimais.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-@Controller
-public class RegisterController {
-
-    @GetMapping("/pet/register")
-    public String novo() {
-        return "/pessoa/cadastroUsu";
-    }
-
-     
-}
